==> 1_classifier.ipynb <==
import pandas as pd

from sklearn.manifold import TSNE
from sklearn.svm import SVC, OneClassSVM
from sklearn.multiclass import OneVsRestClassifier

from sklearn.metrics import average_precision_score
from sklearn.metrics import precision_recall_curve
from sklearn.metrics import precision_score, recall_score

from sklearn.decomposition import NMF
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import SGDClassifier

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import KFold
from sklearn.neural_network import MLPClassifier

seen_df = pd.read_csv('data/0a_labelled_documents.csv')
seen_df.head()

unseen_df = pd.read_csv('data/0b_unlabelled_documents.csv')

df = (pd.concat([seen_df,unseen_df])
      .sort_values('id')
      .sample(frac=1, random_state=1)
      .reset_index(drop=True)
     )

seen_index = df[df['seen']==1].index
print(df.shape)
df.head()

vec = TfidfVectorizer(
    ngram_range=(1,2),
    min_df=4, max_df=0.8, strip_accents='unicode', 
    max_features=10000,
    use_idf=1,
    smooth_idf=1, sublinear_tf=1,
    stop_words="english"
)
vec.fit(df.loc[seen_index,'content'])
ab_X = vec.transform(df['content'])

clf = SVC(kernel='rbf',class_weight='balanced',probability=True)
y = df['relevant']
clf.fit(ab_X[seen_index],y[seen_index])

unseen_index = df[df['seen']==0].index
y_pred = clf.predict_proba(ab_X[unseen_index])

import numpy as np
np.save("data/predictions",y_pred)

df['prediction'] = df['relevant']
df.loc[unseen_index,"prediction"] = y_pred[:,1]

rel_index = df[df['relevant']==1].index
pred_rel_index = df[(df['prediction']>=0.5) & (df['prediction']<1)].index

clf = OneVsRestClassifier(SVC(kernel='linear', class_weight="balanced", probability=True))
classes = ["1 - Mitigation","1 - Adaptation", "1 - Impact"]
y = np.matrix(df[classes])

clf.fit(ab_X[rel_index],y[rel_index]) #fit on the labeled as relevant data
m_y_pred = clf.predict_proba(ab_X[pred_rel_index]) #get probability scores for those predicted to be relevant

np.save("data/m_predictions",m_y_pred)
