==> 2_topic_model.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import sys, os, django

sys.path.append('/home/max/software/django-tmv/tmv_mcc-apsis/BasicBrowser')
os.environ.setdefault("DJANGO_SETTINGS_MODULE", "BasicBrowser.settings")
django.setup()

from tmv_app.models import *

django.db.connection.close()

# Get the topic-term scores associated with the run_id from our database
run_id = 2927 
tts = pd.DataFrame(TopicTerm.objects.filter(run_id=run_id).values('topic_id','term_id','score')).sort_values('topic_id')
tts.head()

# Put them into a matrix, filling empty cells with 0
H = (tts
     .pivot(index="topic_id",columns="term_id")
     .fillna(0).values
    )
H.shape

# Get the vocabulary
vocab = pd.DataFrame(Term.objects.filter(id__in=tts['term_id'].unique()).values('title','id'))
vocab

# Load the data we want to transform into the old model space
df = pd.read_csv('data/included_docs.csv')

# Get the docs with more than 10 words, and strip out copyright boilerplate
df['n_words'] = df['content'].str.findall(r'(\w+)').str.len()
docs = df[df['n_words']>10]
docs['content'] = [x.split("Copyright (C)")[0] for x in docs['content']]
docs['content'] = [re.split("\([C-c]\) [1-2][0-9]{3} Elsevier", x)[0] for x in docs['content']]
docs['content'] = [x.split("Published by Elsevier")[0] for x in docs['content']]
docs['content'] = [x.split("Copyright. (C)")[0] for x in docs['content']]
docs['content'] = [re.split("\. \(C\) [1-2][0-9]{3} ", x)[0] for x in docs['content']]
docs['content'] = [re.split("\. \(C\) Copyright", x)[0] for x in docs['content']]

ids = docs['id']
abstracts = docs['content']

from nltk.stem import SnowballStemmer
import nltk
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import string

def tokenize(text):
    transtable = {ord(c): None for c in string.punctuation + string.digits}
    tokens = nltk.word_tokenize(text.translate(transtable))
    tokens = [i for i in tokens if len(i) > 2 and len(i) < 100]
    return tokens

class snowball_stemmer(object):
    def __init__(self):
        self.stemmer = SnowballStemmer("english")
    def __call__(self, doc):
        return [self.stemmer.stem(t) for t in tokenize(doc)]

tokenizer = snowball_stemmer()
stoplist = set(nltk.corpus.stopwords.words("english"))

# Vectorize the data into a document-term matrix tfidf
vectorizer = TfidfVectorizer(
    max_df=stat.max_df,
    min_df=stat.min_freq,
    ngram_range=(stat.ngram,stat.ngram),
    tokenizer=tokenizer,
    stop_words=stoplist
)
tfidf = vectorizer.fit_transform(abstracts).todense()
vocab_list = vectorizer.get_feature_names()

# Initialise an empty matrix with the vocab in the original topic model
X = np.matrix(np.zeros((len(docs),H.shape[1])))

# Fill this matrix with our data from the new documents
vocab_translate = {row['title']: i for i, row in vocab.iterrows()}
for i, word in enumerate(vocab_list):
    try:
        term_idx = vocab_translate[word]
    except KeyError as e:
        error = f"couldn't find {e}"
    X[:,term_idx] = tfidf[:,i]
    
X.shape

from sklearn.decomposition import NMF

# Initialise an NMF classifier with the topic-terms from the run before 
clf = NMF(
    n_components=ttm.shape[0],
    init="custom",
    max_iter=5
)
clf.components_ = H
clf.n_components_ = H.shape[0]

# Transform the new data into the topic model space
W = clf.transform(
    X
)

# Save the nonzero elements in a long csv

dtm = pd.DataFrame(W)
dtm.columns = tts.topic_id.unique()
dtm.index = ids
dtm = (dtm
       .reset_index()
       .rename(columns={"id":"doc_id"})
       .melt(id_vars="doc_id", var_name="topic_id",value_name="score")
       .query("score>0")
      )
dtm.to_csv('data/doc_topic_scores.csv',index=False)
dtm
