==> 3b_fix_summarise_places.ipynb <==
import pandas as pd
from operator import itemgetter
import ast
import numpy as np


seen_df = pd.read_csv('data/0a_labelled_documents.csv')
unseen_df = pd.read_csv('data/0b_unlabelled_documents.csv')

df = (pd.concat([seen_df,unseen_df])
      .sort_values('id')
      .sample(frac=1, random_state=1)
      .reset_index(drop=True)
     )

seen_index = df[df['seen']==1].index
unseen_index = df[df['seen']==0].index
print(df.shape)

# Get a list of predictions for the unseen documents
y_pred = np.load('data/predictions.npy')
df['prediction'] = df['relevant']
df.loc[unseen_index,"prediction"] = y_pred[:,1]

df = df[df['prediction']>0.5]
print(df.shape)

df.head()

places = pd.read_csv('data/places.csv')
print(len(places.doc_id.unique()))

places = places[places['country_conf']>0.4]

print(len(places.doc_id.unique()))

gm_docs = df.loc[
    (df['title'].str.lower().str.contains("gulf of mexico")) |
    (df['content'].str.lower().str.contains("gulf of mexico")),
    "id"
]
geocolumns = ["word", "country_conf", "feature_code","lat","lon","place_name","feature_class","geonameid"]
gm = pd.DataFrame({"doc_id": gm_docs})
gm[geocolumns] = ["Gulf of Mexico",0.8,"GULF", 25, -90, "Gulf of Mexico", "H", 3523271]


lab_docs = df.loc[
    (df['title'].str.lower().str.contains("labrador sea")) |
    (df['content'].str.lower().str.contains("labrador sea")),
    "id"
]
geocolumns = ["word", "country_conf", "feature_code","lat","lon","place_name","feature_class","geonameid"]
lab = pd.DataFrame({"doc_id": lab_docs})
lab[geocolumns] = ["Labrador Sea",0.8,"SEA", 57, -55, "Labrador Sea", "H", 3424929]

baf_docs = df.loc[
    (df['title'].str.lower().str.contains("baffin bay")) |
    (df['content'].str.lower().str.contains("baffin bay")),
    "id"
]
geocolumns = ["word", "country_conf", "feature_code","lat","lon","place_name","feature_class","geonameid"]
baf = pd.DataFrame({"doc_id": baf_docs})
baf[geocolumns] = ["Baffin Bay",0.8,"BAY", 74, -68, "Baffin Bay", "H", 3831554]

ok_docs = df.loc[
    (df['title'].str.lower().str.contains("sea of okhotsk")) |
    (df['content'].str.lower().str.contains("sea of okhotsk")),
    "id"
]
geocolumns = ["word", "country_conf", "feature_code","lat","lon","place_name","feature_class","geonameid"]
ok = pd.DataFrame({"doc_id": baf_docs})
ok[geocolumns] = ["Sea of Okhotsk",0.8, "SEA", 55, 150, "Sea of Okhotsk", "H", 2127380]

places = pd.concat([places, gm, lab, baf, ok])

places.sort_values('country_conf').head()

geocolumns = ["feature_code","lat","lon","place_name","feature_class","geonameid"]
places.loc[places["word"]=="Pakistan", geocolumns]=["PCLI",30,70,"Islamic Republic of Pakistan","A",1168579]
places.loc[places["word"]=="Colombia", geocolumns]=["PCLI",4,-73.25,"Colombia","A",3686110]
places.loc[places["word"]=="Argentina", geocolumns]=["PCLI",-34,-64,"Argentine Republic","A",3865483]
places.loc[places["word"]=="Sahara", geocolumns] = ["DSRT", 26, 13, "Sahara", "T", 2212709]
places.loc[places["word"]=="Alps",geocolumns] = ["MTS", 46.41667, 10, "Alps", "T", 2661786]
places.loc[places["word"]=="Mediterranean Sea",geocolumns] = ["SEA", 35, 20, "Mediterranean Sea", "T", 2661786]
places.loc[places["word"]=="MEDITERRANEAN",geocolumns] = ["SEA", 35, 20, "Mediterranean Sea", "T", 2661786]
places.loc[places["word"]=="East China",geocolumns] = ["PCLI", 35, 105, "China", "A", 1814991]
places.loc[places["word"]=="South China",geocolumns] = ["PCLI", 35, 105, "China", "A", 1814991]
places.loc[places["word"]=="Great Lakes",geocolumns] = ["LK", 45.68751, -84.43753, "Great Lakes", "H", 4994594]
places.loc[places["word"]=="Catalonia",geocolumns] = ["ADM1", 41.82046, 1.86768, "Catalunya", "A", 3336901]
places.loc[places["word"]=="South Pacific",geocolumns] = ["OCN", -45, -130, "South Pacific Ocean", "H", 4030483]
places.loc[places["word"]=="Gulf Coast",geocolumns] = ["AREA", 29.36901, -95.00565, "Gulf Coast", "L", 7287689]
places.loc[places["word"]=="Gulf coast",geocolumns] = ["AREA", 29.36901, -95.00565, "Gulf Coast", "L", 7287689]
places.loc[places["word"]=="Hainan Island",geocolumns] = ["ISL", 19.2, 109.7, "Hainan Dao", "T", 1809055]
places.loc[places["word"]=="Red Sea",geocolumns] = ["SEA", 20.26735, 38.53455, "Red Sea", "H", 350155]
places.loc[places["word"]=="Himalayan",geocolumns] = ["MTS", 28,84, "Himalayas", "T", 1252558]
places.loc[places["word"]=="Himalayas",geocolumns] = ["MTS", 28,84, "Himalayas", "T", 1252558]
places.loc[places["word"]=="North America's",geocolumns] = ["CONT", 46.07323, -100.54688, "North America", "L", 6255149]
places.loc[places["word"]=="Atlantic Ocean",geocolumns] = ["OCN", 10, -25, "Atlantic Ocean", "H", 3373405]
places.loc[places["word"]=="Scandinavia",geocolumns] = ["RGN", 63, 12, "Scandinavia", "L", 2614165]
places.loc[places["word"]=="California (USA",geocolumns] = ["ADM1", 37.25022, -119.75126, "California", "A", 5332921]
places.loc[places["word"]=="California, USA",geocolumns] = ["ADM1", 37.25022, -119.75126, "California", "A", 5332921]
places.loc[places["word"]=="North Pacific",geocolumns] = ["OCN", 30, -170, "North Pacific Ocean", "H", 4030875]
places.loc[places["word"]=="Huai",geocolumns] = ["STM", 33.133333, 118.5, "Huai He", "H", 1807690]
places.loc[places["word"]=="Washington, DC",geocolumns] = ["PPLC", 38.89511, -77.03637, "Washington", "P", 4140963]
places.loc[places["word"]=="Messinian",geocolumns] = ["ADM2", 37.25, -21.83333, "Nomos Messinias", "A", 257149]
places.loc[places["word"]=="Ionian Sea",geocolumns] = ["SEA", 39, 19, "Ionian Sea", "H", 2463713]
places.loc[places["word"]=="NYC",geocolumns] = ["PPL", 40.71427, -74.00597, "New York City", "P", 5128581]
places.loc[places["word"]=="Indian Ocean",geocolumns] = ["OCN", -10, 70, "Indian Ocean", "P", 1545739]
places.loc[places["word"]=="North Sea",geocolumns] = ["SEA", 55, 3, "North Sea", "P", 2960848]
places.loc[places["word"]=="Philippine Sea",geocolumns] = ["SEA", 20, 135, "Philippine Sea", "P", 1818190]
places.loc[places["word"]=="Black Sea",geocolumns] = ["SEA", 43, 34, "Black Sea", "H", 630673]
places.loc[places["word"]=="Coral Sea",geocolumns] = ["SEA", -20, 155, "Coral Sea", "H", 2194166]
places.loc[places["word"]=="Timor Sea",geocolumns] = ["SEA", -11, 127, "Timor Sea", "H", 2078065]
places.loc[places["word"]=="Hudson Bay",geocolumns] = ["BAY", 60, -85, "Hudson Bay", "H", 5978134]
places.loc[places["word"]=="Bering Sea",geocolumns] = ["SEA", 60, -175, "Bering Sea", "H", 4031788]
places.loc[places["word"]=="Okhotsk Sea",geocolumns] = ["SEA", 55, 150, "Sea of Okhotsk", "H", 2127380]

places.loc[places["place_name"]=="Central Upper Nile",geocolumns] = ["ADM1", 10, 32.7, "Upper Nile", "A", 381229]


places.loc[places["place_name"]=="Gobolka Woqooyi Galbeed","place_name"] = "Woqooyi Galbeed"

places = places[places["place_name"]!="Pacific County"]
places = places.loc[places["word"]!="B.V."]
places = places[places["word"]!="MMT"]
places = places[places["word"]!="Yellow"]
places = places[places["word"]!="Hadley"]
places = places[places["word"]!="Western North"]
places = places[places["word"]!="colonies"]
places = places[places["word"]!="TN"]
places = places[places["word"]!="NH"]
places = places[places["word"]!="Mn"]
places = places[places["word"]!="Tx"]
places = places[places["word"]!="TX"]
places = places[places["word"]!="Tn"]
places = places[places["word"]!="FL"]
places = places[places["word"]!="Spartina"]
places = places[places["word"]!="Tamarix"]
places = places[places["word"]!="Eurasia"]
places = places[places["word"]!="Phillyrea"]
places = places[places["word"]!="N-15"]
places = places[places["word"]!="LT50"]
places = places[places["word"]!="POSEIDON"]
places = places[places["word"]!="LC50"]
places = places[places["word"]!="El Nio"]
places = places[places["word"]!="La Nia"]
places = places[places["word"]!="Red"]
places = places[places["word"]!="Gulf Stream"]
places = places[places["word"].str.len()>2]
places = places[places["word"]!="NH 1"]
places = places[places["word"]!="Quercus"]
places = places[(places["word"]!="ZJP")]


places.lat = places.lat.astype(float)
places.lon = places.lon.astype(float)

places = places.drop_duplicates(["doc_id","word"])

print(places.shape)

places.to_csv("data/clean_place_df.csv",index=False)


import pycountry_convert as pc
def get_cont(x):
    continents = {
        'NA': 'North America',
        'SA': 'South America', 
        'AS': 'Asia',
        'OC': 'Oceania',
        'AF': 'Africa',
        'EU': 'Europe'
    }
    try:
        return continents[pc.country_alpha2_to_continent_code(pc.country_alpha3_to_country_alpha2(x['country_predicted']))]
    except:
        return None
places['continent'] = places.apply(lambda x: get_cont(x), axis=1)
other_country_continents = [
    ("ESH","Africa"),
    ("TLS","Oceania"),
    ("XKX","Europe")
    
]
for country,continent in other_country_continents:
    places.loc[places['country_predicted']==country,"continent"] = continent
    
print(places.shape)

places.to_csv("data/clean_place_df.csv",index=False)
    

places.head()

pubs = [
    ("S. Karger AG, Basel",["Basel","Switzerland"]),
    ("Licensee MDPI, Basel",["Basel","Switzerland"])
]
for p, place in pubs:
    pub_docs = df.loc[df.content.str.contains(p)]
    remove = places[(places.doc_id.isin(pub_docs['id'])) & (places['place_name'].isin(place))]
    print(remove.shape[0])
    places = places.drop(index=remove.index)

    
print(places.shape)
places.to_csv("data/clean_place_df.csv",index=False)
places.head()

places[places['doc_id']==612295]
