==> 3c_get_institution_locations.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import sys, os, django

sys.path.append('/home/max/software/django-tmv/tmv_mcc-apsis/BasicBrowser')
os.environ.setdefault("DJANGO_SETTINGS_MODULE", "BasicBrowser.settings")
django.setup()

from scoping.models import *

django.db.connection.close()

# Get a list of predictions for the unseen documents
y_pred = np.load('data/predictions.npy')

# Load the documents and add the predictions
seen_df = pd.read_csv('data/0a_labelled_documents.csv')
unseen_df = pd.read_csv('data/0b_unlabelled_documents.csv')

df = (pd.concat([seen_df,unseen_df])
      .sort_values('id')
      .sample(frac=1, random_state=1)
      .reset_index(drop=True)
     )

seen_index = df[df['seen']==1].index
unseen_index = df[df['seen']==0].index

df['prediction'] = df['relevant']
df.loc[unseen_index,"prediction"] = y_pred[:,1]
print(df.shape)

# Get just likely relevant documents
df = df[df['prediction']>0.33]
print(df.shape)
df.head()

dais = pd.DataFrame.from_dict(DocAuthInst.objects.filter(doc__id__in=df['id'],institution__iregex='\w').values())
dais.head()

import pycountry

def clean_country(x):
    country = x.split(',')[-1]
    if "USA" in country:
        country = "USA"
    return country.strip('. ]"\'')
        
dais['country'] = dais['institution'].apply(lambda x: clean_country(x))
dais.loc[dais['country']=='CA 95616','country'] = "USA"
dais.loc[dais['country']=="Peoples R China",'country'] = "China"
dais.loc[dais['country']=="Rep of Georgia",'country'] = "Georgia"
dais.loc[dais['country']=="South Korea",'country'] = "Republic of Korea"
dais.loc[dais['country']=="Trinidad Tobago",'country'] = "Trinidad and Tobago"
dais.loc[dais['country']=="Trinid & Tobago",'country'] = "Trinidad and Tobago"
dais.loc[dais['country']=="North Ireland","country"] = "Northern Ireland"  
dais.loc[dais['country']=="DEM REP CONGO", "country"] = "Congo, The Democratic Republic of the"
dais.loc[dais['country']=="Mongol Peo Rep", "country"] = "Mongolia"
dais.loc[dais['country']=="Laos", "country"] = "Lao"
dais.loc[dais['country']=="Bosnia & Herceg", "country"] = "Bosnia"
dais.loc[dais['country']=="U Arab Emirates", "country"] = "United Arab Emirates"
dais.loc[dais['country']=="Cote Ivoire", "country"] = "Côte d'Ivoire"
dais.loc[dais['country']=="Fr Polynesia", "country"] = "French Polynesia"
dais.loc[dais['country']=="Papua N Guinea", "country"] = "Papua New Guinea"
dais.loc[dais['country']=="Rep Congo", "country"] = "Republic of the Congo"
dais.loc[dais['country']=="Fr Polynesia", "country"] = "French Polynesia"
dais.loc[dais['country']=="Swaziland", "country"] = "Eswatini"
dais.loc[dais['country']=="Neth Antilles", "country"] = "Netherlands Antilles"
dais.loc[dais['country']=="Guinea Bissau", "country"] = "Guinea-Bissau"
dais.loc[dais['country']=="Cape Verde", "country"] = "Cabo Verde"
dais.loc[dais['country']=="Cent Afr Republ", "country"] = "Central African Republic"
dais.loc[dais['country']=="Equat Guinea", "country"] = "Equatorial Guinea"
dais.loc[dais['country']=="Cent Afr Republ", "country"] = "Central African Republic"
dais.loc[dais['country']=="Cent Afr Republ", "country"] = "Central African Republic"
dais.loc[dais['country']=="Cent Afr Republ", "country"] = "Central African Republic"



dais['country'] = dais['country'].str.replace('St','Saint')
dais['country'] = dais['country'].str.replace('&','and')

for c in dais.country.unique():
    try:
        country = pycountry.countries.search_fuzzy(c)[0]
    except:
        country = pycountry.historic_countries.get(name=c)
        if country is None:
            print("Could not find it!") 
            break
    dais.loc[dais['country']==c,["alpha_3","name","alpha_2"]] = [country.alpha_3, country.name, country.alpha_2]
#list(dais.country.unique())

dais.to_csv('data/institutional_locations.csv',index=False)

dais[dais['country']=='CA 95616'].institution.unique()
