==> 6_summary_data.ipynb <==
import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import cartopy.crs as ccrs

who_regions = pd.read_csv('data/who-regions.csv')
who_regions.head()

df = pd.read_csv('data/included_docs.csv').rename(columns={"prediction": "0 - relevance - prediction"})
df.head()

places = pd.read_csv('data/clean_place_df.csv')
if "WHO region" not in places.columns:
    places = places.merge(who_regions.rename(columns={"Code": "country_code3"})[["country_code3","WHO region"]])
places.head()

dais = pd.read_csv('data/institutional_locations.csv')
dais = (dais
        .rename(columns={"alpha_3": "country_code3"})
        .merge(who_regions.rename(columns={"Code": "country_code3"})[["country_code3","WHO region"]])
       )

auth_strings = dais.drop_duplicates(['AU','doc_id']).sort_values('position').groupby('doc_id')['AU'].agg(
    lambda x: "; ".join(x)
).reset_index().rename(columns={"doc_id": "id", "AU": "authors"})

dais.head()

auth_strings

if "authors" not in df.columns:
    df = df.merge(auth_strings)

pred_cols = [x for x in df.columns if "pred" in x]
df = df[['id','content','title','PY','authors'] + pred_cols]
df.to_csv('data/included_studies.csv',index=False)
df.head()

with pd.ExcelWriter('data/institutions_authors.xlsx') as writer: 

    # For each who region
    for i, who in enumerate(who_regions['WHO region'].unique()):
        print(who)
        # get the papers mentioning a place in this region
        dids = set(places[places['WHO region']==who].doc_id)

        top_institutions = (dais[dais.doc_id.isin(dids)]
                            .groupby('institution')['doc_id']
                            .nunique()
                            .sort_values(ascending=False)
                            .to_frame("n_studies")
                            .reset_index()
                           )
        top_institutions.to_excel(writer, sheet_name=f"Institutions about {who}", index=False)
        
        top_authors = (dais[dais.doc_id.isin(dids)]
                            .groupby('AU')['doc_id']
                            .nunique()
                            .sort_values(ascending=False)
                            .to_frame("n_studies")
                            .reset_index()
                           )
        top_authors.to_excel(writer, sheet_name=f"Authors about {who}", index=False)
        
        # Get the papers with an affiliation from this region
        dids = set(dais[dais['WHO region']==who].doc_id)

        top_institutions = (dais[dais.doc_id.isin(dids)]
                            .groupby('institution')['doc_id']
                            .nunique()
                            .sort_values(ascending=False)
                            .to_frame("n_studies")
                            .reset_index()
                           )
        top_institutions.to_excel(writer, sheet_name=f"Institutions from {who}", index=False)
        
        top_authors = (dais[dais.doc_id.isin(dids)]
                            .groupby('AU')['doc_id']
                            .nunique()
                            .sort_values(ascending=False)
                            .to_frame("n_studies")
                            .reset_index()
                           )
        top_authors.to_excel(writer, sheet_name=f"Authors from {who}", index=False)
